Initial airshots tracking + other various tracking
Airshot tracking works :) I manually watched the logged ticks for a
few demos and the current heuristic (in the air for at least 16 ticks
before being killed) seems reasonable.

Supstats2 defines airshots based off the victim's height off the
ground:
https://github.com/F2/F2s-sourcemod-plugins/blob/f893dd7e49c8f09c70a86c0024c31ee303b3d1b8/supstats2/supstats2.sp#L1135

But that would be really hard for us to implement... I don't think
either heuristic is perfect but we can iterate on it, perhaps
incorporating player velocity and animation state.